Rename cellulariot to libhat
