+++ deps/libhat/src/Adafruit_ADS1015.c
