+++ deps/libhat/src/MMA8452Q.h
