+++ deps/libhat/src/libhat.c
@@ -14,7 +14,7 @@ Date                  Author                  Modification
 #include "SDL_Pi_HDC1000.h"
 #include "MMA8452Q.h"
 #include "Adafruit_ADS1015.h"
-#include "cellulariot.h"
+#include "libhat.h"
 
 #ifndef NULL
 #define NULL 0
