+++ deps/libhat/src/libhat.h
